Add note for 'RandCropByPosNegLabeld'
It seems that the transform randomly selects a foreground point from image, then use it as center crop. This means that it can find the closest voxel that is just outside the SC and use it as the center (hence it includes the SC).
Source: https://github.com/Project-MONAI/MONAI/issues/452#issuecomment-636065502